refactor(net): move biqueue module to own file
